Basic item state and placement pass
Finished the fist basic pass at machine and item interactions. As well as adding a handler to manage and update items models to match their data using a scriptable object.

Player can now pick up a portafilter and place it into the holder of the grinder. This interactions will only work if the player is holding a portafilter in their left hand. I will add the right hand once all of the remaining components are implemented.